pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "gifts-sales"
version = "0.1.0"
description = "Telegram collectible gifts marketplace userbot"
requires-python = ">=3.12"
dependencies = [
    "telethon>=1.43.2",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.19",
    "typer>=0.12",
    "rich>=13.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.0",
    "ruff>=0.4",
]

[project.scripts]
gifts-sales = "app.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
